feat: decompose to partials
